Move investment signals to separate module
