+++ datahub/investment/models.py
@@ -4,8 +4,6 @@ import uuid
 
 from django.conf import settings
 from django.db import models, transaction
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from model_utils import Choices
 
 from datahub.core.constants import InvestmentProjectStage
@@ -322,19 +320,3 @@ class IProjectDocument(BaseModel, ArchivableModel):
             investment_doc.save()
 
         return investment_doc
-
-
-@receiver(post_save, sender=InvestmentProject, dispatch_uid='project_post_save')
-def project_post_save(sender, **kwargs):
-    """Creates a project code for investment projects on creation.
-
-    Projects with a CDMS project code do not get a new project code.
-
-    This generates project codes for fixtures loaded via manage.py loaddata
-    (i.e. when kwargs['raw'] is True), though that may need to change if
-    fixed project codes are required for that fixtures.
-    """
-    instance = kwargs['instance']
-    created = kwargs['created']
-    if created and not instance.cdms_project_code:
-        InvestmentProjectCode.objects.create(project=instance)
